updated calculation

+++ src/main/java/percentages/percentagesReducer.java
@@ -1,7 +1,6 @@
 package percentages;
 
 import org.apache.hadoop.io.DoubleWritable;
-import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Reducer;
 
@@ -9,7 +8,7 @@ import java.io.IOException;
 
 public class percentagesReducer extends Reducer<Text, DoubleWritable, Text, DoubleWritable> {
     public void reduce(Text key, Iterable<DoubleWritable> values, Context context) throws IOException, InterruptedException {
-        int sum = 0;
+        double sum = 0;
         for (DoubleWritable i : values) {
             sum += i.get();
         }
